task 1

+++ 0x01-variables_if_else_while/1-last_digit.c
@@ -13,7 +13,7 @@ int main(void)
 
 	srand(time(0));
 	n = rand() - RAND_MAX / 2;
-	int num = n % 10;
+	long num = n % 10;
 
 	if (num > 5)
 		printf("Last digit of %i is %i and is greater than 5\n", n, num);
